requirements.txt
streamlit
openai
sounddevice
numpy
wave
requests
python-dotenv
